Move Sanitizer service to gaphor.UML package
