+++ gaphor/UML/sanitizerservice.py
