+++ gaphor/UML/tests/conftest.py
@@ -1,0 +1,1 @@
+from gaphor.diagram.tests.fixtures import diagram, element_factory, event_manager
