+++ pyproject.toml
@@ -80,7 +80,7 @@ gaphorconvert = 'gaphor.tools.gaphorconvert:main'
 "export_menu" = "gaphor.ui.menufragment:MenuFragment"
 "tools_menu" = "gaphor.ui.menufragment:MenuFragment"
 "copy" = "gaphor.services.copyservice:CopyService"
-"sanitizer" = "gaphor.services.sanitizerservice:SanitizerService"
+"sanitizer" = "gaphor.UML.sanitizerservice:SanitizerService"
 "namespace" = "gaphor.ui.namespace:Namespace"
 "toolbox" = "gaphor.ui.toolbox:Toolbox"
 "diagrams" = "gaphor.ui.mainwindow:Diagrams"
